Working on create/delete handlers

+++ src/main/java/software/amazon/lightsail/cfn/instance/DeleteHandler.java
@@ -3,11 +3,15 @@ package software.amazon.lightsail.cfn.instance;
 import com.amazonaws.services.lightsail.AmazonLightsail;
 import com.amazonaws.services.lightsail.AmazonLightsailClientBuilder;
 import com.amazonaws.services.lightsail.model.DeleteInstanceRequest;
+import com.amazonaws.services.lightsail.model.GetInstanceRequest;
+import com.amazonaws.services.lightsail.model.GetInstanceResult;
+import com.amazonaws.services.lightsail.model.Instance;
 import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
 import software.amazon.cloudformation.proxy.Logger;
 import software.amazon.cloudformation.proxy.ProgressEvent;
-import software.amazon.cloudformation.proxy.OperationStatus;
 import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
+
+import java.time.Duration;
 
 public class DeleteHandler extends BaseHandler<CallbackContext> {
 
@@ -22,13 +26,64 @@ public class DeleteHandler extends BaseHandler<CallbackContext> {
 
         final ResourceModel model = request.getDesiredResourceState();
 
+        final CallbackContext currentContext = callbackContext == null
+                ? CallbackContext.builder()
+                .stabilizationRetriesRemaining(Constants.NUMBER_OF_STATUS_POLL_RETRIES)
+                .build()
+                : callbackContext;
+
+        if (currentContext.getStabilizationRetriesRemaining() == 0) {
+            throw new RuntimeException(Constants.DELETE_TIMED_OUT_MESSAGE);
+        }
+
+        if(currentContext.getStatus() != null && currentContext.getStatus().equals(Constants.STATUS_DELETE_PENDING)) {
+            return checkDeleteStatus(proxy, request, currentContext, logger);
+        }
+
         DeleteInstanceRequest deleteInstanceRequest = new DeleteInstanceRequest();
         deleteInstanceRequest.setInstanceName(model.getInstanceName());
         proxy.injectCredentialsAndInvoke(deleteInstanceRequest, lightsailClient::deleteInstance);
 
-        return ProgressEvent.<ResourceModel, CallbackContext>builder()
-            .resourceModel(model)
-            .status(OperationStatus.SUCCESS)
-            .build();
+        CallbackContext newCallbackContext = CallbackContext.builder()
+                .stabilizationRetriesRemaining(currentContext.getStabilizationRetriesRemaining() - 1)
+                .status(Constants.STATUS_DELETE_PENDING)
+                .build();
+
+        return ProgressEvent.defaultInProgressHandler(
+                newCallbackContext,
+                (int) Duration.ofSeconds(Constants.CALLBACK_DELAY_SECONDS).getSeconds(),
+                model
+        );
+    }
+
+    private ProgressEvent<ResourceModel, CallbackContext> checkDeleteStatus(
+            final AmazonWebServicesClientProxy proxy,
+            final ResourceHandlerRequest<ResourceModel> request,
+            final CallbackContext callbackContext,
+            final Logger logger) {
+
+        final ResourceModel model = request.getDesiredResourceState();
+
+        GetInstanceRequest getInstanceRequest = new GetInstanceRequest();
+        getInstanceRequest.setInstanceName(model.getInstanceName());
+
+        GetInstanceResult getInstanceResult =
+                proxy.injectCredentialsAndInvoke(getInstanceRequest, lightsailClient::getInstance);
+        Instance instance = getInstanceResult.getInstance();
+        if(instance == null) {
+            logger.log(String.format("Instance %s has been deleted", model.getInstanceName()));
+            return ProgressEvent.defaultSuccessHandler(model);
+        }
+
+        logger.log(String.format("Still deleting instance %s", model.getInstanceName()));
+        logger.log(instance.toString());
+        CallbackContext newCallbackContext = CallbackContext.builder()
+                .stabilizationRetriesRemaining(callbackContext.getStabilizationRetriesRemaining() - 1)
+                .status(Constants.STATUS_DELETE_PENDING)
+                .build();
+        return ProgressEvent.defaultInProgressHandler(
+                newCallbackContext,
+                (int) Duration.ofSeconds(Constants.CALLBACK_DELAY_SECONDS).getSeconds(),
+                model);
     }
 }
